requirements.txt
absl-py==1.0.0
actionlib==1.12.1
angles==1.9.12
appdirs==1.4.4
beautifulsoup4==4.10.0
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.24
certifi==2021.5.30
charset-normalizer==2.0.7
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cv-bridge==1.13.0
cycler==0.10.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
distro==1.6.0
docutils==0.18.1
dynamic-reconfigure==1.6.3
easydict==1.9
filelock==3.6.0
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gdown==4.3.1
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
idna==3.3
image-geometry==1.13.0
imageio==2.9.0
interactive-markers==1.11.5
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
kiwisolver==1.3.2
kornia==0.6.4
laser-geometry==1.6.7
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.4.3
message-filters==1.14.12
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.3
pandas==1.2.5
Pillow==9.0.1
prettytable==2.4.0
protobuf==3.19.1
pycuda==2021.1
pyfakewebcam==0.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytools==2021.2.9
pytorch-quantization==2.1.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
resource-retriever==1.12.6
rosbag==1.14.12
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.12
roslaunch==1.14.12
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.12
rosmake==1.14.9
rosmaster==1.14.12
rosmsg==1.14.12
rosnode==1.14.12
rosparam==1.14.12
rospkg==1.3.0
rospy==1.14.12
rosservice==1.14.12
rostest==1.14.12
rostopic==1.14.12
rosunit==1.14.9
roswtf==1.14.12
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.21
scikit-image==0.18.3
scipy==1.7.1
segmentation-refinement==0.4
sensor-msgs==1.12.8
six==1.16.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
soupsieve==2.3.1
sphinx-glpi-theme==0.3
tabulate==0.8.9
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
thop==0.0.31.post2005241907
tifffile==2021.10.12
timm==0.4.12
topic-tools==1.14.12
torch==1.8.2+cu111
torch2trt==0.3.0
torchaudio==0.8.2
torchvision==0.9.2+cu111
tqdm==4.62.3
typing-extensions==3.10.0.2
urdfdom-py==0.4.6
urllib3==1.26.7
wcwidth==0.2.5
xacro==1.13.14
